CMakeLists.txt: Allow to turn off the install target
For embeddeding the library into existing CMake projects it is generally
not required, and wanted, to execute the installation target for the
library as well when then installation target of the existing CMake
project is called.

Allow users to turn off the installation target with JSON_INSTALL.
For backwards compatibility it is ON by default.

Signed-off-by: Thomas Braun <thomas.braun@byte-physics.de>